Moved to 'Current/Reflection' + top-level namespace `current`
